requirements.txt
streamlit
langchain
langchain_core
langchain_community
python-dotenv
langchain-groq